Small macro class re-organization
