+++ lib/dry/schema/macros.rb
@@ -1,2 +1,6 @@
-require 'dry/schema/macros/required'
+require 'dry/schema/macros/each'
+require 'dry/schema/macros/filled'
+require 'dry/schema/macros/hash'
+require 'dry/schema/macros/maybe'
 require 'dry/schema/macros/optional'
+require 'dry/schema/macros/required'
